financial_management_app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
av==13.1.0
blinker==1.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
geopandas==1.0.1
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
livekit==0.18.1
livekit-agents==0.11.3
livekit-api==0.8.0
livekit-plugins-openai==0.10.7
livekit-protocol==0.7.0
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.1.0
numpy==2.0.2
openai==1.55.0
packaging==24.1
pandas==2.2.3
pillow==10.3.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==5.9.8
pydantic==2.10.0
pydantic_core==2.27.0
PyJWT==2.10.1
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
watchfiles==0.24.0
Werkzeug==3.1.3
yarl==1.18.0
